requirements.txt
absl-py==2.1.0
alabaster==0.7.16
annotated-types==0.7.0
app-model==0.3.1
appdirs==1.4.4
asciitree==0.3.3
asttokens==2.4.1
attrs==25.1.0
babel==2.17.0
blinker==1.9.0
build==1.2.2.post1
cachey==0.2.1
cairocffi==1.7.1
CairoSVG==2.8.2
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.3.0
cssselect2==0.8.0
cycler==0.12.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dask==2024.8.0
debugpy==1.8.8
decorator==5.1.1
defusedxml==0.7.1
docstring_parser==0.16
docutils==0.21.2
einops==0.8.0
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.1.0
fasteners==0.19
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.0.3
flexcache==0.3
flexparser==0.4
fonttools==4.55.3
freetype-py==2.5.1
fsspec==2024.10.0
grpcio==1.70.0
h5py==3.12.1
HeapDict==1.0.1
hsluv==5.0.4
huggingface-hub==0.29.3
idna==3.10
imageio==2.36.0
imagesize==1.4.1
importlib_metadata==8.5.0
importlib_resources==6.5.2
in-n-out==0.2.1
ipykernel==6.29.5
ipython==8.18.1
ipywidgets==8.1.5
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.7
kornia==0.7.4
kornia_rs==0.1.7
lazy_loader==0.4
llvmlite==0.43.0
locket==1.0.0
magicgui==0.10.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
napari==0.5.6
napari-console==0.1.3
napari-plugin-engine==0.2.0
napari-plugin-manager==0.1.4
napari-svg==0.2.1
narwhals==1.26.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
nibabel==5.3.2
npe2==0.7.8
numba==0.60.0
numcodecs==0.12.1
numpy==1.24.1
numpydoc==1.8.0
open3d==0.19.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pandas_flavor==0.7.0
parso==0.8.4
partd==1.4.2
PartSegCore_compiled_backend==0.15.12
patsy==1.0.1
pillow==11.1.0
pingouin==0.5.5
Pint==0.24.4
platformdirs==4.3.6
plotly==6.0.0
pooch==1.8.2
prompt_toolkit==3.0.48
protobuf==5.29.3
psutil==6.1.0
psygnal==0.12.0
pure_eval==0.2.3
pyconify==0.2.1
pycparser==2.22
pydantic==2.11.2
pydantic-compat==0.1.2
pydantic_core==2.33.1
pydicom==2.4.4
Pygments==2.18.0
PyOpenGL==3.1.9
pyparsing==3.2.1
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.8.1
PyQt6-Qt6==6.8.2
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
pytz==2024.2
# pywin32==308  # Windows-specific package, skipped for Linux containers
PyYAML==6.0.2
pyzmq==26.2.0
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==14.0.0
rpds-py==0.22.3
safetensors==0.5.3
scikit-image==0.24.0
scikit-learn==1.6.0
scipy==1.13.1
seaborn==0.13.2
shellingham==1.5.4
SimpleITK==2.4.1
six==1.17.0
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
stack-data==0.6.3
statsmodels==0.14.4
superqt==0.7.3
sympy==1.13.3
tabulate==0.9.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.5.0
tifffile==2024.8.30
timm==1.0.15
tinycss2==1.4.0
tomli==2.2.1
tomli_w==1.2.0
toolz==1.0.0
# torch and torchvision are installed separately in Dockerfile for GPU/CPU flexibility
tornado==6.4.1
tqdm==4.67.0
traitlets==5.14.3
triangle==20250106
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
vispy==0.14.3
wcwidth==0.2.13
webencodings==0.5.1
Werkzeug==3.0.6
widgetsnbextension==4.0.13
wrapt==1.17.2
xarray==2024.7.0
zarr==2.18.2
zipp==3.21.0
